@@ -3,19 +3,19 @@ apply plugin: 'idea'
 apply plugin: 'maven-publish'
 
 buildscript {
-    ext.android_buildtools_version = '27.0.3'
-    ext.android_support_version = '26.1.0'
+    ext.android_buildtools_version = '28.0.3'
+    ext.androidx_version = '1.0.0'
     ext.jackson_version = '2.8.6'
     ext.jpa_version = '2.1.1'
-    ext.kotlin_version = '1.2.20'
-    ext.rxjava2_version = '2.1.9'
-    ext.reactor_version = '3.1.4.RELEASE'
+    ext.kotlin_version = '1.3.11'
+    ext.rxjava2_version = '2.2.3'
+    ext.reactor_version = '3.2.2.RELEASE'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.2'
+        classpath 'com.android.tools.build:gradle:3.2.1'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "org.jetbrains.dokka:dokka-gradle-plugin:0.9.8"
