@@ -10,26 +10,23 @@ apply plugin: 'kotlin-android'
 apply plugin: 'maven-publish'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     buildToolsVersion "$android_buildtools_version"
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles 'proguard-rules.pro'
+        multiDexEnabled true
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
     }
     lintOptions {
         abortOnError false
@@ -41,16 +38,17 @@ android {
 }
 
 dependencies {
-    compile project(':requery')
-    compile "com.android.support:support-v4:$android_support_version"
-    compile "com.android.support:recyclerview-v7:$android_support_version"
-    compile 'net.zetetic:android-database-sqlcipher:3.5.9'
-    compile 'io.requery:sqlite-android:3.21.0'
-    testCompile 'junit:junit:4.12'
-    androidTestCompile project(':requery-test')
-    androidTestCompile 'com.android.support.test:runner:1.0.2'
-    androidTestCompile 'com.android.support.test:rules:1.0.2'
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation project(':requery')
+    implementation 'androidx.sqlite:sqlite:2.0.0'
+    implementation "androidx.recyclerview:recyclerview:$androidx_version"
+    implementation 'net.zetetic:android-database-sqlcipher:4.0.1'
+    implementation 'io.requery:sqlite-android:3.25.3'
+    testImplementation 'junit:junit:4.12'
+    androidTestImplementation project(':requery-test')
+    androidTestImplementation 'androidx.test:runner:1.1.0'
+    androidTestImplementation 'androidx.test:rules:1.1.0'
+    androidTestImplementation 'androidx.multidex:multidex:2.0.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 }
 
 publish.dependsOn 'assembleRelease'
