@@ -5,19 +5,16 @@ buildscript {
 }
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-
-kapt {
-    generateStubs = true
-}
+apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     buildToolsVersion "$android_buildtools_version"
 
     defaultConfig {
         applicationId "io.requery.android.example.kotlinapp"
         minSdkVersion 15
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -36,13 +33,13 @@ android {
 }
 
 dependencies {
-    compile "com.android.support:appcompat-v7:$android_support_version"
-    compile "com.android.support:recyclerview-v7:$android_support_version"
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    compile "io.reactivex.rxjava2:rxjava:$rxjava2_version"
-    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
-    compile 'io.requery:requery:1.3.0'
-    compile 'io.requery:requery-android:1.3.0'
-    compile 'io.requery:requery-kotlin:1.3.0'
-    kapt 'io.requery:requery-processor:1.3.0'
+    implementation "androidx.appcompat:appcompat:$androidx_version"
+    implementation "androidx.recyclerview:recyclerview:$androidx_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation "io.reactivex.rxjava2:rxjava:$rxjava2_version"
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
+    //implementation project(':requery') // replace with 'io.requery:requery:<version>'
+    implementation project(':requery-android') // replace with 'io.requery:requery-android:<version>'
+    implementation project(':requery-kotlin') // replace with 'io.requery:requery-kotlin:<version>'
+    kapt project(':requery-processor') // replace with 'io.requery:requery-processor:<version>'
 }
