@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     buildToolsVersion "$android_buildtools_version"
 
     defaultConfig {
         applicationId "io.requery.android.example.app"
         minSdkVersion 15
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -16,10 +16,6 @@ android {
             minifyEnabled true
             proguardFiles 'proguard-rules.pro'
         }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
     }
     lintOptions {
         abortOnError false
@@ -30,11 +26,12 @@ android {
 }
 
 dependencies {
-    compile "com.android.support:appcompat-v7:$android_support_version"
-    compile "com.android.support:recyclerview-v7:$android_support_version"
-    compile "io.reactivex.rxjava2:rxjava:$rxjava2_version"
-    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
-    compile 'io.requery:requery:1.3.0'
-    compile 'io.requery:requery-android:1.3.0'
-    annotationProcessor 'io.requery:requery-processor:1.3.0'
+    implementation "androidx.appcompat:appcompat:$androidx_version"
+    implementation "androidx.recyclerview:recyclerview:$androidx_version"
+    implementation "io.reactivex.rxjava2:rxjava:$rxjava2_version"
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
+    implementation project(':requery') // replace with 'io.requery:requery:<version>'
+    implementation project(':requery-android') // replace with 'io.requery:requery-android:<version>'
+    implementation 'androidx.multidex:multidex:2.0.0'
+    annotationProcessor project(':requery-processor') // replace with 'io.requery:requery-processor:<version>'
 }
