@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 requery.io
+ * Copyright 2018 requery.io
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,6 +19,7 @@ package io.requery.sql.platform;
 import io.requery.query.Expression;
 import io.requery.query.element.LimitedElement;
 import io.requery.query.element.OrderByElement;
+import io.requery.query.element.QueryElement;
 import io.requery.sql.GeneratedColumnDefinition;
 import io.requery.sql.IdentityColumnDefinition;
 import io.requery.sql.Mapping;
@@ -26,9 +27,11 @@ import io.requery.sql.Platform;
 import io.requery.sql.UserVersionColumnDefinition;
 import io.requery.sql.VersionColumnDefinition;
 import io.requery.sql.gen.Generator;
+import io.requery.sql.gen.InsertGenerator;
 import io.requery.sql.gen.LimitGenerator;
 import io.requery.sql.gen.OffsetFetchGenerator;
 import io.requery.sql.gen.OrderByGenerator;
+import io.requery.sql.gen.UpdateGenerator;
 import io.requery.sql.gen.UpsertMergeGenerator;
 
 import java.util.Map;
@@ -39,17 +42,21 @@ import java.util.Map;
 public class Generic implements Platform {
 
     private final GeneratedColumnDefinition generatedColumnDefinition;
-    private final LimitGenerator limitDefinition;
+    private final LimitGenerator limitGenerator;
     private final VersionColumnDefinition versionColumnDefinition;
-    private final Generator<Map<Expression<?>, Object>> upsertDefinition;
-    private final Generator<OrderByElement> orderByDefinition;
+    private final Generator<QueryElement<?>> insertGenerator;
+    private final Generator<Map<Expression<?>, Object>> updateGenerator;
+    private final Generator<Map<Expression<?>, Object>> upsertGenerator;
+    private final Generator<OrderByElement> orderByGenerator;
 
     public Generic() {
         generatedColumnDefinition = new IdentityColumnDefinition();
-        limitDefinition = new OffsetFetchGenerator();
+        limitGenerator = new OffsetFetchGenerator();
         versionColumnDefinition = new UserVersionColumnDefinition();
-        upsertDefinition = new UpsertMergeGenerator();
-        orderByDefinition = new OrderByGenerator();
+        insertGenerator = new InsertGenerator();
+        updateGenerator = new UpdateGenerator();
+        upsertGenerator = new UpsertMergeGenerator();
+        orderByGenerator = new OrderByGenerator();
     }
 
     @Override
@@ -98,7 +105,7 @@ public class Generic implements Platform {
 
     @Override
     public Generator<LimitedElement> limitGenerator() {
-        return limitDefinition;
+        return limitGenerator;
     }
 
     @Override
@@ -107,13 +114,23 @@ public class Generic implements Platform {
     }
 
     @Override
+    public Generator<QueryElement<?>> insertGenerator() {
+        return insertGenerator;
+    }
+
+    @Override
+    public Generator<Map<Expression<?>, Object>> updateGenerator() {
+        return updateGenerator;
+    }
+
+    @Override
     public Generator<Map<Expression<?>, Object>> upsertGenerator() {
-        return upsertDefinition;
+        return upsertGenerator;
     }
 
     @Override
     public Generator<OrderByElement> orderByGenerator() {
-        return orderByDefinition;
+        return orderByGenerator;
     }
 
     @Override
